move into the jvm only test folder
