+++ zio-schema-json/shared/src/test/scala-2/zio/schema/codec/JsonCodecSpec.scala
@@ -1,6 +1,5 @@
 package zio.schema.codec
 
-import java.nio.charset.StandardCharsets
 import java.time.{ ZoneId, ZoneOffset }
 
 import scala.collection.immutable.ListMap
@@ -9,7 +8,7 @@ import zio.Console._
 import zio._
 import zio.json.JsonDecoder.JsonError
 import zio.json.ast.Json
-import zio.json.{ DeriveJsonEncoder, JsonEncoder, JsonStreamDelimiter }
+import zio.json.{ DeriveJsonEncoder, JsonEncoder }
 import zio.schema.CaseSet._
 import zio.schema._
 import zio.schema.annotation._
@@ -18,7 +17,7 @@ import zio.schema.codec.JsonCodec.JsonEncoder.charSequenceToByteChunk
 import zio.schema.codec.JsonCodecSpec.PaymentMethod.{ CreditCard, PayPal, WireTransfer }
 import zio.schema.codec.JsonCodecSpec.Subscription.{ OneTime, Recurring }
 import zio.schema.meta.MetaSchema
-import zio.stream.{ ZPipeline, ZStream }
+import zio.stream.ZStream
 import zio.test.Assertion._
 import zio.test.TestAspect._
 import zio.test._
@@ -603,33 +602,7 @@ object JsonCodecSpec extends ZIOSpecDefault {
           WithComplexOptionField(Some(Order(1, BigDecimal.valueOf(10), "test"))),
           charSequenceToByteChunk("""{"order":{"id":1,"value":10,"description":"test"}}""")
         )
-      },
-      test("case class with empty option field is decoded by stream") {
-        val names     = Gen.option(Gen.elements("John", "Jane", "Jermaine", "Jasmine"))
-        val age       = Gen.option(Gen.int(1, 99))
-        val person    = names.zipWith(age)(AllOptionalFields(_, _, None))
-        val delimiter = Gen.elements(JsonStreamDelimiter.Array, JsonStreamDelimiter.Newline)
-        val codec     = JsonCodec.jsonEncoder(AllOptionalFields.schema)
-
-        check(Gen.chunkOfBounded(0, 3)(person), delimiter) {
-          (people, delim) =>
-            val indent        = if (delim == JsonStreamDelimiter.Array) Some(1) else None
-            val encodedPeople = people.map(p => codec.encodeJson(p, indent))
-            val encoded = delim match {
-              case JsonStreamDelimiter.Array =>
-                encodedPeople.mkString("[", ",", "]")
-              case JsonStreamDelimiter.Newline =>
-                encodedPeople.mkString("", "\n", "\n")
-            }
-
-            assertDecodesJsonStream(
-              AllOptionalFields.schema,
-              people,
-              charSequenceToByteChunk(encoded),
-              delim
-            )
-        }
-      } @@ TestAspect.jvmOnly
+      }
     ),
     suite("enums")(
       test("case name aliases - default") {
@@ -1495,21 +1468,6 @@ object JsonCodecSpec extends ZIOSpecDefault {
   ) = {
     val result = ZStream.fromChunk(chunk).via(JsonCodec.schemaBasedBinaryCodec[A](cfg)(schema).streamDecoder).runCollect
     assertZIO(result)(equalTo(Chunk(value)))
-  }
-
-  private def assertDecodesJsonStream[A](
-    schema: Schema[A],
-    value: Chunk[A],
-    chunk: Chunk[Byte],
-    delimiter: JsonStreamDelimiter
-  ) = {
-    val result = ZStream
-      .fromChunk(chunk)
-      .via(ZPipeline.decodeCharsWith(StandardCharsets.UTF_8))
-      .via(JsonCodec.jsonDecoder(schema).decodeJsonPipeline(delimiter))
-      .runCollect
-      .either
-    assertZIO(result)(isRight(equalTo(value)))
   }
 
   private def assertEncodesThenDecodesFallback[A, B](
